test: migrate `from_native` tests
